Add revive XP system

+++ GTF_Xp/Extensions/Information/NetworkingInfo/StaticXpInfo.cs
@@ -4,16 +4,16 @@ namespace GTFuckingXP.Information.NetworkingInfo
 {
     public struct StaticXpInfo : IXpData
     {
-        public StaticXpInfo(uint xpGain, uint debuffXp, int levelScaling, Vector3 position)
+        public StaticXpInfo(int xpGain, int debuffXp, int levelScaling, Vector3 position)
         {
-            XpGain = xpGain;
-            DebuffXp = debuffXp;
+            XpGain = (int) xpGain;
+            DebuffXp = (int) debuffXp;
             LevelScalingXpDecrese = levelScaling;
             Position = position;
         }
 
-        public uint XpGain { get; set; }
-        public uint DebuffXp { get; set; }
+        public int XpGain { get; set; }
+        public int DebuffXp { get; set; }
         public int LevelScalingXpDecrese { get; set; }
         public Vector3 Position { get; set; }
     }
